Create question JSON output URI
I wanted to be able to quickly look at the current JSON
representation of any given question, so added a URI to
display the JSON for any given question. I also tweaked
the system to pull the CSS for the AsciiDoctor pages
from the /static area in the DropWizard server.