Complete the upload ZIP file feature
This feature allows a clean system to have a ZIP of
previously downloaded Recipes and Questions uploaded
into the database. It clears both collections and
loads them from the ZIP file. It also resets the max
sequence number for the questions, or creates it
if loading into a fresh database.

Other small cleanups and scripts are included in
this commit as well.