Guard against empty answer questions/category change
There are still a few empty-answer questions in the database
which must not make it into any quiz. Add guard code to avoid
that, as well as handle empty-answer questions on the show
question page.

Also, allow the editing of Category, but if Category is changed
clear all the attributes to insure that only attributes which
match the category are included in the question.

Finally, allow the selecting of questions from a DIFFERENT
category into a given quiz. This is because some questions
are valid for various different quizzes.