Cleanups and enhancements
Remove some unused fields from the Question POJO. Make it such
that links to questions off of the test script output page are
launched in new tabs. Clean up the list of attributes being
output on the test script output page. Ensure that attributes
are always sorted when output (and, honestly, no reason to not
have them sorted in the DB just to make scanning them visually
a bit easier when that is required).