Change how deleted is handled
If a question is quarintined or deployed, no longer return an
error but simply mark it as deleted. Then during cleanup it can
be finally deleted if no longer deployed (though that cleanup
behavior is not implemented in this commit).

Also, be more inclusive in testing and rejecting inappropriate
questions. The list of reasons to reject was incomplete.