Continue work on "KNOWLEDGE" (was "GENERAL") category
Switch to using KNOWLEDGE rather than GENERAL, and change the
category reports to more accurately be attribute reports. Create
a category-level report so we can report out all the new KNOWLEDGE
questions as they are created.

Enhance the DSL to allow category sub-selection on the getAll and
getAny type commands.