Quiz record report
Add a report to show all quiz records. Additionally, add the feature
of recording the last person to generate a key for a given quiz. All
retrievals before the 4.0 deployment will not have this field set,
but in time the older quiz records will be expunged by the background
housekeeping job.